requirements.txt
asgiref==3.2.7
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-static==0.0.6
Django==2.2.1
django-extensions==2.2.9
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-filter==0.2.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
phonenumbers==8.10.8
Pillow==5.4.1
postgres==2.2.2
psycopg2-binary==2.7.7
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
